uclibc: fix tolower and locales
The function towlower doesn't work with locales different from C.
Issue was introduced in commit: 8cde3a9bf2856dcb9a759dec7ecb04a68e712254

Call to setlocale is needed for correct generation of the table uplow_diff.
Otherwise you receive compile time error "range assumption error" after
uncommenting the call.

Similar problem described here:
http://lists.uclibc.org/pipermail/uclibc/2015-March/048852.html

This commit fix the problem by using int32_t values.

+++ uclibc/lib/contrib/uclibc/extra/locale/locale_mmap.h
@@ -45,7 +45,7 @@ typedef struct {
 #ifdef __WCHAR_ENABLED
 	const unsigned char tblwctype[__LOCALE_DATA_WCctype_TBL_LEN];
 	const unsigned char tblwuplow[__LOCALE_DATA_WCuplow_TBL_LEN];
-	const int16_t tblwuplow_diff[__LOCALE_DATA_WCuplow_diff_TBL_LEN];
+	const int32_t tblwuplow_diff[__LOCALE_DATA_WCuplow_diff_TBL_LEN];
 /* 	const unsigned char tblwcomb[WCcomb_TBL_LEN]; */
 	/* width?? */
 #endif
